[lab7] add sfml canvas implementation; refactor console canvas

+++ lab7/Slides/Canvas/ICanvas.h
@@ -28,8 +28,8 @@ public:
 	virtual ~ICanvas() = default;
 
 	virtual void DrawLine(const Point& from, const Point& to) = 0;
-	virtual void DrawEllipse(const Point& leftTop, float width, float height) = 0;
-	virtual void FillEllipse(const Point& leftTop, float width, float height) = 0;
+	virtual void DrawEllipse(const Point& leftTop, double width, double height) = 0;
+	virtual void FillEllipse(const Point& leftTop, double width, double height) = 0;
 	virtual void FillPolygon(std::vector<Point>&) = 0;
 	virtual void SetFillColor(uint32_t) = 0;
 	virtual void SetLineColor(uint32_t) = 0;
